Prototype
Description (written by ChatGPT):  

This program creates a simulation of a creature with a segmented body and multiple legs. The body consists of a series of circular segments connected by lines. Each segment has the ability to move and follow the position of either the mouse cursor (in the case of the first segment) or the position of the segment that precedes it in the chain. The program also includes the ability to add and remove legs from each segment.

Each segment contains two legs, which are drawn as lines with a circular foot at the end. The legs also move with the segments they are attached to, and the length of each leg is limited to prevent them from extending too far. The legs move based on a trigonometric calculation that takes into account the rotation of the segment they are attached to.

The program uses the p5.js library for drawing and animation, and all of the code is written in JavaScript.